Move carousel data fetching into openlibrary/core
Extract the carousel family (get_carousel_data, generic_carousel,
get_ia_carousel_books, format_book_data, presets/config) out of the home
page controller into a new canonical openlibrary.core.carousels module, and
move caching_prethread into openlibrary.utils.request_context. The homepage
and loans page now both build carousels from openlibrary.core: this removes
the upstream->openlibrary-plugin import in account.py and the pre-existing
core->plugin inversions in lending.py and fulltext.py. The config map is
flattened to name->subject, and an explicit empty carousel list now yields
no carousels instead of silently falling back to all of them.
