+++ openlibrary/core/fulltext.py
@@ -5,8 +5,8 @@ from urllib.parse import urlencode
 import httpx
 
 from infogami import config
+from openlibrary.core.carousels import format_book_data
 from openlibrary.core.lending import get_availability_async
-from openlibrary.plugins.openlibrary.home import format_book_data
 from openlibrary.utils.async_utils import async_bridge
 from openlibrary.utils.request_context import req_context, site
