+++ openlibrary/plugins/upstream/account.py
@@ -40,6 +40,7 @@ from openlibrary.core.auth import ExpiredTokenError, HMACToken, MissingKeyError
 from openlibrary.core.auth import TimedOneTimePassword as OTP
 from openlibrary.core.booknotes import Booknotes
 from openlibrary.core.bookshelves import Bookshelves
+from openlibrary.core.carousels import get_carousel_data
 from openlibrary.core.db import get_db
 from openlibrary.core.follows import PubSub
 from openlibrary.core.lending import get_loan_history_data
@@ -1278,8 +1279,6 @@ class account_loans(delegate.page):
         docs = get_loans_of_user(user.key)
         loan_history_data = get_loan_history_data(username, page=page)
         featured_subjects = get_cached_featured_subjects()
-
-        from openlibrary.plugins.openlibrary.home import get_carousel_data
 
         staff_picks = get_carousel_data(carousels=("staff_picks",))["staff_picks"]
